requirements.txt
einops==0.4.1
gdown==4.6.4
numpy==1.24.0
opencv-python==4.6.0.66
pandas==1.5.3
Pillow==9.4.0
requests==2.25.1
scikit-image==0.20.0
scipy==1.8.1
seaborn==0.10.1
setuptools==59.6.0
timm==0.6.12
torch-dct==0.1.5
torchattacks==3.3.0
tqdm==4.47.0
torchmetrics==0.11.4
IQA-pytorch==0.1
adversarial-robustness-toolbox==1.15.0
